fix(dashboard): extract /jobs* BFF routes to bff_jobs.py (file_length gate)
The staging merge (new /jobs/cancel route) + this branch's SSE stream routes
pushed bff.py to 321 SLOC, over the 300 file_length cap (CI red). Extract all
six /jobs* handlers (list, launch, steer, cancel, stream-token, stream) into
`routes/bff_jobs.py::register_jobs_routes`, mirroring the existing
register_agent_routes / register_admin_routes pattern. Verbatim move — routes
register on the same router at the same point, behaviour unchanged.

bff.py 321 → 224 SLOC.
